python/requirements.txt
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pydantic==2.8.2
loguru==0.7.2
